refactor: the registration blocks filtering
